User and Email changes
* User/Admin Crud apis done
* sendEmail functionality added

+++ controllers/courses.js
@@ -51,6 +51,7 @@ export const getCourse = asyncHandler(async (req, res, next) => {
  */
 export const addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -58,6 +59,15 @@ export const addCourse = asyncHandler(async (req, res, next) => {
     return next(
       ErrorResponse(`No bootcamp with id of ${req.params.bootcampId}`),
       404
+    );
+  }
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
     );
   }
 
@@ -81,6 +91,15 @@ export const updateCourse = asyncHandler(async (req, res, next) => {
     return next(ErrorResponse(`No course with id of ${req.params.id}`), 404);
   }
 
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
+
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -102,6 +121,15 @@ export const deleteCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     return next(ErrorResponse(`No course with id of ${req.params.id}`), 404);
+  }
+
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
   }
 
   await Course.findByIdAndRemove(req.params.id);
